+++ @types/apis/psn.d.ts
@@ -1,0 +1,6 @@
+type getUserTitlesOptions = {
+  limit?: number;
+  offset?: number;
+};
+
+export type { getUserTitlesOptions };