+++ CHANGELOG.md
@@ -8,5 +8,6 @@
 - Add instructions for authenticaticating with PSN
 - Add method to retrieve PSN API authorisation object using authentication credentials
 - Add method to get an PSN account ID from a username
+- Add ability to retrieve a list of games for which a user has gained trophies
 
 [unreleased]: https://github.com/yndajas/video-game-library/compare/9b793f6...HEAD